Modified it a bit to make it work better.
Also I added the resource files since they're going to be needed by the final version once I get it done with it. tomorrow I'll add the pause feature and also make it so that upon fail if the difficulty is set to hard to prompt them to try it on easy.

Also I'll warn the user if their FPS is greater than 60 to try toggling the FPS limiter to be only 60.